Dev ah branch

+++ EcoNest/EcoNest/Views/Home/Cart/CartViewModel.swift
@@ -9,100 +9,94 @@ import SwiftUI
 import FirebaseAuth
 import FirebaseFirestore
 
-/// ViewModel responsible for managing cart-related operations such as adding to and fetching from Firestore.
+/// ViewModel responsible for managing cart-related operations.
 class CartViewModel: ObservableObject {
     
-    /// List of cart items currently stored in Firestore for the user.
+    // Published properties to notify SwiftUI views of changes
     @Published var cartProducts: [Cart] = []
     @Published var isLoading = false
     @Published var selectedDate: Date = Date()
-    
-    private var cartListener: ListenerRegistration?
 
+    private var cartListener: ListenerRegistration? // Listener for real-time cart updates
+    private var authListener: AuthStateDidChangeListenerHandle? // Listener for auth state changes
+
+    
+    /// Initializes the view model and sets up a Firebase Auth listener
     init() {
-        fetchCartData()
-    }
+        // Add a listener to monitor authentication state changes
+        authListener = Auth.auth().addStateDidChangeListener { [weak self] auth, user in
+            // Use [weak self] to prevent retain cycles
+            guard let self = self else { return }
 
-    deinit {
-        cartListener?.remove()
-    }
-    
-    /// Fetches all cart items from Firestore and resolves them into complete Cart objects with full Product data.
-    func fetchCartData() {
-        
-        isLoading = true
-        cartListener?.remove() // Remove previous one
-        
-        let db = FirebaseManager.shared.firestore
-        
-        guard let userId = Auth.auth().currentUser?.uid else {
-            print("User not logged in")
-            isLoading = false
-            return
+            // If the user logs out, clear cart data and remove cart listener
+            if user == nil {
+                self.cartProducts = []
+                self.cartListener?.remove()
+            } else {
+                self.fetchCartData() // Fetch cart data from Firestore for the logged-in user
+            }
         }
-        
-        cartListener = db.collection("users")
-            .document(userId)
-            .collection("cart")
-            .addSnapshotListener { snapshot, error in
-                
-                guard let documents = snapshot?.documents else {
-                    print("No cart data found.")
-                    self.isLoading = false
-                    return
-                }
+    }
 
-                var fetchedCart: [Cart] = []
-                let group = DispatchGroup()
+    
+    /// Fetches the user's cart data from Firestore and listens for real-time updates.
+    func fetchCartData() {
+        isLoading = true
+        cartListener?.remove()
 
-                for doc in documents {
-                    let data = doc.data()
-                    
-                    guard let productId = data["productId"] as? String,
-                          let quantity = data["quantity"] as? Int,
-                          let price = data["price"] as? Double else { continue }
+        let db = FirebaseManager.shared.firestore
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
 
-                    group.enter()
+        cartListener = userDoc.collection("cart").addSnapshotListener { snapshot, error in
+            guard let documents = snapshot?.documents else {
+                print("No cart data found.")
+                self.isLoading = false
+                return
+            }
 
-                    db.collection("product").document(productId).getDocument { productDoc, error in
-                        defer { group.leave() }
+            var fetchedCart: [Cart] = []
+            let group = DispatchGroup()
 
-                        if let productData = productDoc?.data() {
-                            let product = Product(
-                                id: productDoc!.documentID,
-                                name: productData["name"] as? String ?? "",
-                                description: productData["description"] as? String ?? "",
-                                price: productData["price"] as? Double ?? 0.0,
-                                image: productData["image"] as? String ?? "",
-                                quantity: productData["quantity"] as? Int ?? 0,
-                                careLevel: productData["careLevel"] as? String ?? "",
-                                size: productData["size"] as? String ?? ""
-                            )
+            for doc in documents {
+                let data = doc.data()
 
+                guard let productId = data["productId"] as? String,
+                      let quantity = data["quantity"] as? Int,
+                      let price = data["price"] as? Double else { continue }
+
+                group.enter()
+
+                db.collection("product").document(productId).getDocument { productDoc, error in
+                    defer { group.leave() }
+
+                    if let productDoc = productDoc, productDoc.exists {
+                        let product = try? productDoc.data(as: Product.self)
+
+                        if let product = product {
                             let cartItem = Cart(
                                 id: doc.documentID,
                                 product: product,
                                 quantity: quantity,
                                 price: price
                             )
-
-                           
                             fetchedCart.append(cartItem)
-                            
                         }
                     }
                 }
-
-                group.notify(queue: .main) {
-                    self.cartProducts = fetchedCart
-                    self.isLoading = false
-                }
             }
-    }
 
+            group.notify(queue: .main) {
+                self.cartProducts = fetchedCart
+                self.isLoading = false
+            }
+        }
+    }
+    
+    
     /// Calculates the total price of all items in the cart.
     /// - Returns: The sum of (quantity × unit price) for each cart item.
     func calculateTotal() -> Double {
+        
         var price: Double = 0
         // Iterate through each cart item
         cartProducts.forEach { cartItem in
@@ -112,48 +106,16 @@ class CartViewModel: ObservableObject {
         return price
     }
     
-    /// Remove items from the user's cart in Firestore and updates the local cart state.
-    /// - Parameter index: An IndexSet containing the indices of items to remove.
-    func removeFormCart(index: IndexSet) {
-        let db = FirebaseManager.shared.firestore
-        guard let userId = Auth.auth().currentUser?.uid else {
-            print("User must be logged in to add to cart.")
-            return
-        }
-        guard let i = index.first else { return }
-        
-        // Get the corresponding cart item using the index
-        let cartItem = cartProducts[i]
-        
-        // Reference the Firestore document for the cart item and attempt deletion
-        db.collection("users")
-            .document(userId)
-            .collection("cart")
-            .document(cartItem.id)
-            .delete { err in
-                
-                if let err = err {
-                    // Print the error if deletion fails
-                    print("Failed to delete from cart: \(err.localizedDescription)")
-                    return
-                }
-                
-                // If deletion is successful, remove the item from the local cart list on the main thread
-                DispatchQueue.main.async {
-                    self.cartProducts.remove(at: i)
-                }
-            }
-    }
     
     /// Adds a new order to the user's "orders" collection in Firestore.
     /// - Parameter locationId: The ID of the selected pickup location.
     func addOrder(locationId: String) {
-        let db = FirebaseManager.shared.firestore
-        guard let userId = Auth.auth().currentUser?.uid else {
-            print("User must be logged in to add to cart.")
-            return
-        }
-        let cartData = cartProducts.map { cart in
+        
+        // Ensure user is logged in before proceeding
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
+        
+        // Converts the product items into a dictionary format for Firestore.
+        let productData = cartProducts.map { cart in
             return [
                 "name": cart.product.name ?? "",
                 "price": cart.price,
@@ -163,64 +125,83 @@ class CartViewModel: ObservableObject {
             ]
         }
 
-        db.collection("users")
-            .document(userId)
+        // Add the new order to the user's orders collection
+        userDoc
             .collection("orders")
             .document()
             .setData([
-                "products": cartData,
+                "products": productData,
                 "total": calculateTotal(),
                 "date": Timestamp(date: selectedDate),
                 "pickupLocation": locationId,
-                "status" : OrderStatus.awaitingPickup.rawValue
+                "status": OrderStatus.awaitingPickup.rawValue
             ]) { error in
                 if let error = error {
                     print("Failed to place order: \(error.localizedDescription)")
                     return
                 }
+                
+                // Update inventory quantities for ordered products
+                self.reduceQuantity()
 
-                // Reduce product quantities
-                for cart in self.cartProducts {
-                    let productRef = db.collection("product").document(cart.product.id ?? "")
+                // Clear cart after order placement and inventory update
+                self.clearCartFromFirestore()
+            }
+    }
+    
+    
+    /// Updates product inventory in Firestore by reducing quantity based on the user's cart.
+    private func reduceQuantity() {
+        
+        let db = FirebaseManager.shared.firestore
+        
+        for cart in self.cartProducts {
+            
+            let productCollection = db.collection("product").document(cart.product.id ?? "")
+            
+            // Fetch the current product document
+            productCollection.getDocument { document, error in
+                if let document = document, document.exists,
+                   let data = document.data(),
+                   let currentQuantity = data["quantity"] as? Int {
                     
-                    productRef.getDocument { document, error in
-                        if let document = document, document.exists,
-                           let data = document.data(),
-                           let currentQuantity = data["quantity"] as? Int {
-                            
-                            let newQuantity = max(currentQuantity - cart.quantity, 0)
-                            
-                            productRef.updateData(["quantity": newQuantity]) { error in
-                                if let error = error {
-                                    print("Failed to update product quantity: \(error.localizedDescription)")
-                                }
-                            }
+                    // Calculate the new quantity, ensuring it doesn't go below 0
+                    let newQuantity = max(currentQuantity - cart.quantity, 0)
+                    
+                    // Update Firestore with the new quantity
+                    productCollection.updateData(["quantity": newQuantity]) { error in
+                        if let error = error {
+                            print("Failed to update product quantity: \(error.localizedDescription)")
                         }
                     }
                 }
-
-                // Clear cart after reducing inventory
-                self.clearCartFromFirestore(userId: userId)
             }
+        }
     }
 
     
-    private func clearCartFromFirestore(userId: String) {
-        let db = FirebaseManager.shared.firestore
+    /// Clears the current user's cart from Firestore and updates the local cart state.
+    private func clearCartFromFirestore() {
         
-        db.collection("users")
-            .document(userId)
+        // Get the document reference for the current user; exit if not available
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
+        
+        // Access the "cart" subcollection for the current user
+        userDoc
             .collection("cart")
             .getDocuments { snapshot, error in
+                // Check if documents were retrieved successfully
                 guard let documents = snapshot?.documents else {
                     print("No cart items found.")
                     return
                 }
                 
+                // Iterate through each document (cart item) and delete it
                 for doc in documents {
                     doc.reference.delete()
                 }
                 
+                // Clear the local cart state on the main thread
                 DispatchQueue.main.async {
                     self.cartProducts.removeAll() // Clear local cart as well
                 }
@@ -228,27 +209,29 @@ class CartViewModel: ObservableObject {
                 print("Cart cleared after order placement.")
             }
     }
+
     
-    func increaseQuantity(cart: Cart, change: upQuantity) {
+    /// Updates the quantity of a specific cart item in both local state and Firestore.
+    /// - Parameters:
+    ///   - cart: The cart item to be updated.
+    ///   - change: A Boolean indicating whether to increment (true) or decrement (false) the quantity.
+    func updateQuantity(cart: Cart, change: Bool) {
         
+        // Find the index of the cart item in the local array; exit if not found
         guard let index = cartProducts.firstIndex(where: { $0.id == cart.id }) else { return }
         
-        guard let userId = Auth.auth().currentUser?.uid else {
-            print("User must be logged in to add to cart.")
-            return
-        }
-        if change == .increase {
+        // Get the document reference for the current user; exit if not available
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
+        
+        // Adjust the quantity locally based on the change flag
+        if change {
             cartProducts[index].quantity += 1
         } else {
             cartProducts[index].quantity -= 1
         }
         
-
-        // Update in Firestore
-        let db = FirebaseManager.shared.firestore
-        
-        db.collection("users")
-            .document(userId)
+        // Update the corresponding cart document in Firestore with the new quantity
+        userDoc
             .collection("cart")
             .document(cart.id)
             .updateData([
@@ -260,14 +243,47 @@ class CartViewModel: ObservableObject {
             }
     }
 
-    func removeCartItem(cart: Cart) {
-        let db = FirebaseManager.shared.firestore
-        guard let userId = Auth.auth().currentUser?.uid else {
-            print("User must be logged in to add to cart.")
+
+    /// Removes an item from the user's cart in Firestore and updates the local cart state.
+    /// - Parameter index: An IndexSet indices of items to remove.
+    func removeFormCart(index: IndexSet) {
+        
+        // Get the document reference for the current user; exit if not available
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
+        
+        // Safely extract the first index from the IndexSet and check bounds
+        guard let i = index.first, i < cartProducts.count else {
+            print("Invalid index")
             return
         }
-        db.collection("users")
-            .document(userId)
+        
+        // Delete the item from the Firestore cart collection
+        userDoc
+            .collection("cart")
+            .document(cartProducts[i].id) // Retrieve the cart item id to be deleted
+            .delete { err in
+                if let err = err {
+                    print("Failed to delete from cart: \(err.localizedDescription)")
+                    return
+                }
+
+                // Update the local cart on the main thread after successful deletion
+                DispatchQueue.main.async {
+                    self.cartProducts.remove(at: i)
+                }
+            }
+    }
+
+
+    /// Removes a specific cart item from the user's Firestore cart and updates the local cart state.
+    /// - Parameter cart: The cart item to be removed.
+    func removeFormCart(cart: Cart) {
+        
+        // Get the document reference for the current user; exit if not available
+        guard let userDoc = FirebaseManager.shared.getCurrentUser() else { return }
+        
+        // Delete the specific cart item document from the "cart" subcollection in Firestore
+        userDoc
             .collection("cart")
             .document(cart.id)
             .delete { err in
@@ -277,14 +293,10 @@ class CartViewModel: ObservableObject {
                 }
 
                 DispatchQueue.main.async {
+                    // Remove any item with the same ID from the local cart
                     self.cartProducts.removeAll { $0.id == cart.id }
                 }
             }
     }
 
-}
-
-enum upQuantity {
-    case increase
-    case decrease
 }
